peers and talkgroup refactor (#7)
