+++ src/main.rs
@@ -1,7 +1,12 @@
-use dmrpal::{debug, echo, master, sleep, slot, SystemState, Systemstate};
+use dmrpal::{
+    debug, echo, master,
+    peers::{Peer, Peertype},
+    sleep, slot,
+    talkgroups::{Talkgroup, TgActivate},
+};
 use std::collections::{hash_map::HashMap, hash_set::HashSet};
 use std::net::{IpAddr, Ipv4Addr, UdpSocket};
-use std::{io, str, string, time::SystemTime};
+use std::{io, str, time::SystemTime};
 
 mod db;
 mod hb;
@@ -10,13 +15,6 @@ const SOFTWARE_VERSION: u64 = 1;
 const USERACTIVATED_DISCONNECT_TG: u32 = 4000;
 const MY_ID: u32 = 235045402;
 const REMOTE_PEER: &str = "78, 129, 135, 43";
-
-#[derive(Debug, PartialEq)]
-enum Peertype {
-    Local,
-    Friend,
-    All,
-}
 
 #[derive(Debug, PartialEq)]
 enum Masterstate {
@@ -28,266 +26,6 @@ enum Masterstate {
     Connected,
     WaitingPong,
     Logout,
-}
-
-enum TgActivate {
-    Static(u32),
-    Ua(u32),
-}
-
-struct Peer {
-    id: u32,
-    callsign: String,
-    duplex: u8,
-    frequency: String,
-    software: String,
-    latitude: f32,
-    last_check: SystemTime,
-    longitude: f32,
-    power: u16,
-    height: u16,
-    ip: std::net::SocketAddr,
-    talk_groups: HashMap<u32, Talkgroup>,
-    tx_bytes: usize,
-    options: String,
-    peer_type: Peertype,
-    rx_bytes: usize,
-    slot: slot::Slot,
-    tg_expire: u64,
-}
-
-#[derive(Debug)]
-struct Talkgroup {
-    expire: u64,
-    id: u32,
-    la: SystemTime,
-    routeable: Peertype,
-    sl: u8,
-    ua: bool,
-    time_stamp: SystemTime,
-}
-
-impl Peer {
-    fn new() -> Self {
-        Self {
-            id: 0,
-            callsign: string::String::default(),
-            duplex: 0,
-            frequency: string::String::default(),
-            software: string::String::default(),
-            latitude: 0.0,
-            last_check: SystemTime::now(),
-            longitude: 0.0,
-            power: 0,
-            height: 0,
-            ip: std::net::SocketAddr::new(IpAddr::V4(Ipv4Addr::new(0, 0, 0, 0)), 0),
-            talk_groups: HashMap::from([
-                (0, Talkgroup::default()),
-                (31337, Talkgroup::set(2, TgActivate::Static(31337), None)),
-                (2351, Talkgroup::set(1, TgActivate::Static(2351), None)),
-                (235, Talkgroup::set(1, TgActivate::Static(235), None)),
-                (844, Talkgroup::set(2, TgActivate::Static(844), None)),
-                (840, Talkgroup::set(2, TgActivate::Static(840), None)),
-                (123, Talkgroup::set(1, TgActivate::Static(123), None)),
-                (113, Talkgroup::set(1, TgActivate::Static(113), None)),
-                (3, Talkgroup::set(1, TgActivate::Static(3), None)),
-                (2, Talkgroup::set(1, TgActivate::Static(2), None)),
-                (1, Talkgroup::set(1, TgActivate::Static(1), None)),
-            ]),
-            tx_bytes: 0,
-            options: string::String::default(),
-            peer_type: Peertype::Local,
-            rx_bytes: 0,
-            slot: slot::Slot::init(),
-            tg_expire: 0,
-        }
-    }
-
-    // Check if the peer is allowed to sign in.
-    fn acl(&self) -> bool {
-        let known_peers = vec![000000];
-        for k in known_peers {
-            if self.id.eq(&k) {
-                return false;
-            }
-        }
-        true
-    }
-
-    fn connect_master(&mut self) -> Masterstate {
-        let myid = hb::RPTLPacket { id: MY_ID };
-        let pip = std::net::SocketAddr::from(std::net::SocketAddrV4::new(
-            std::net::Ipv4Addr::new(78, 129, 135, 43),
-            55555,
-        ));
-        let mut rx_buff = [0; hb::RX_BUFF_MAX];
-        let mut state = Masterstate::LoginRequest;
-        let sock = match UdpSocket::bind("0.0.0.0:55555") {
-            Ok(s) => s,
-            Err(e) => {
-                eprintln!("There was an error binding: {}", e);
-                std::process::exit(-1);
-            }
-        };
-        loop {
-            sock.send_to(&myid.request_login(), pip).unwrap();
-
-            let (_, src) = match sock.recv_from(&mut rx_buff) {
-                Ok(rs) => (rs),
-
-                Err(e) => {
-                    eprintln!("There was an error listening: {}", e);
-                    std::process::exit(-1);
-                }
-            };
-
-            match &rx_buff[..6] {
-                hb::RPTACK => match &state {
-                    Masterstate::Disable => {}
-                    Masterstate::Disconnected => {}
-                    Masterstate::LoginRequest => {
-                        sock.send_to(&myid.password_response(rx_buff), pip).unwrap();
-                        println!("sending password");
-                        state = Masterstate::LoginPassword;
-                        sleep(80000);
-                    }
-                    Masterstate::LoginPassword => {
-                        sock.send_to(&myid.info(), pip).unwrap();
-                        println!("sending info");
-                        state = Masterstate::Connected;
-                        sleep(80000);
-                    }
-                    Masterstate::Connected => {
-                        sock.send_to(&myid.ping(), pip).unwrap();
-                        println!("connected");
-                        self.ip = src;
-                        break;
-                    }
-                    Masterstate::Options => {}
-                    Masterstate::Logout => {}
-                    _ => {}
-                },
-                hb::RPTNAK => {
-                    println!("MASTER Connect: Received NAK");
-                    state = Masterstate::Disconnected;
-                    break;
-                }
-                _ => println!("MASTER Connect: Packet not handled!\n{:X?}", rx_buff),
-            }
-        }
-        state
-    }
-
-    fn options(&mut self) {
-        for opts in self.options.split(';') {
-            if opts.len() < 3 {
-                continue;
-            }
-            match &opts[..3] {
-                "TS1" | "TS2" => match opts.chars().nth(2) {
-                    Some(s) => {
-                        let slot = s as u8 - 48;
-                        let mut tg: u32 = 0;
-                        for i in opts[5..].bytes() {
-                            if i > 47 && i < 58 {
-                                tg = tg * 10;
-                                tg = tg + i as u32 - 48;
-                            }
-                        }
-                        self.talk_groups
-                            .insert(tg, Talkgroup::set(slot, TgActivate::Static(tg), None));
-                        println!("OPTIONS {}, Added {} {}", self.id, slot, tg);
-                    }
-                    None => continue,
-                },
-                "UAT" => {
-                    let mut t: u64 = 0;
-                    for i in opts[4..].bytes() {
-                        if i > 47 && i < 58 {
-                            t = t * 10;
-                            t = t + i as u64 - 48;
-                        }
-                    }
-                    self.tg_expire = t;
-                }
-                _ => continue,
-            }
-        }
-    }
-
-    // Set the peer ID
-    fn pid(&mut self, buff: &[u8; 4]) {
-        self.id = ((buff[0] as u32) << 24)
-            | ((buff[1] as u32) << 16)
-            | ((buff[2] as u32) << 8)
-            | (buff[3] as u32);
-    }
-}
-
-impl Talkgroup {
-    // return a default value for talkgroup
-    fn default() -> Self {
-        Self {
-            expire: 0,
-            id: 0,
-            la: SystemTime::now(),
-            routeable: Peertype::Local,
-            sl: 1,
-            ua: false,
-            time_stamp: SystemTime::now(),
-        }
-    }
-
-    // Remove a talkgroup from a peer
-    fn ua_clear(&mut self) -> bool {
-        if self.ua {
-            return match self.time_stamp.elapsed() {
-                Ok(ts) => {
-                    if ts.as_secs() > self.expire {
-                        // If the talkgroup has traffic, skip and try again when there's no traffic
-                        if let Ok(la) = self.la.elapsed() {
-                            if la.as_secs() <= 5 {
-                                return true;
-                            }
-                        };
-                        println!("Removing TG: {}, From Peer: {}", self.id, self.id);
-                        false
-                    } else {
-                        true
-                    }
-                }
-                Err(_) => {
-                    println!("There was an error passing time for UA, removing TG!");
-                    false
-                }
-            };
-        }
-        true
-    }
-
-    // Set a talk group to a peer
-    fn set(sl: u8, tg: TgActivate, exp: Option<u64>) -> Self {
-        let (ua, talk_group, expire) = match tg {
-            TgActivate::Static(u) => (false, u, 0),
-            TgActivate::Ua(u) => {
-                let e: u64 = match exp {
-                    Some(v) => v,
-                    None => 900,
-                };
-                (true, u, e)
-            }
-        };
-
-        Self {
-            expire,
-            id: talk_group,
-            la: SystemTime::now(),
-            routeable: Peertype::Local,
-            sl,
-            ua,
-            time_stamp: SystemTime::now(),
-        }
-    }
 }
 
 fn echo(sock: &std::net::UdpSocket, dst: std::net::SocketAddr, data: &Vec<[u8; 55]>) {
@@ -503,7 +241,7 @@ fn main() {
                         }
                     }
                 }
-                
+
                 Masterstate::Options => {
                     let options = hb::RPTOPacket::construct(
                         MY_ID,
